requirements.txt
# Hotel Cleaning Bot - dependencies
aiogram>=3.4.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
